% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{Global_Permutations}
\alias{Global_Permutations}
\title{Perform Global Permutations}
\usage{
Global_Permutations(
  permutationMatrix,
  permut_col,
  cellPropMatrix,
  spotGeneMatrix,
  LigandVectorIndex,
  ReceptorVectorIndex,
  null_expression,
  nBoot
)
}
\arguments{
\item{permutationMatrix}{A matrix containing permutations.}

\item{permut_col}{A column matrix of permutations.}

\item{cellPropMatrix}{A matrix of cell type proportions.}

\item{spotGeneMatrix}{A matrix of gene expressions at spots.}

\item{LigandVectorIndex}{A vector of ligand indices.}

\item{ReceptorVectorIndex}{A vector of receptor indices.}

\item{null_expression}{A matrix of null expression values.}

\item{nBoot}{Number of bootstrap iterations.}
}
\value{
A matrix with the results of the global permutations.
}
\description{
This function performs global permutations on the spatial transcriptomics data.
}
