% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_local.R
\docType{data}
\name{result_local}
\alias{result_local}
\title{result_local data and spatial coordinates for SpaCCI
Local Regional Data}
\format{
A list containing:
\describe{
  \item{dataframelist}{A list of data frames, each containing p-value results for spatial neighborhoods.}
  \item{RegionIDs_matrix}{A list of matrices, each containing the spot IDs for specific neighborhoods.}
}
}
\usage{
data(result_local)
}
\description{
This dataset contains the results of spatial neighborhood analyses.
These objects can be used for testing and running example analyses with the SpaCCI package.
}
\examples{
  data(result_local)
  print(result_local)
}
\keyword{datasets}
