% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_local_spatial_coords_df.R
\docType{data}
\name{result_local_spatial_coords_df}
\alias{result_local_spatial_coords_df}
\title{Local Spatial Coordinates Data Frame}
\format{
A data frame with X rows and 3 columns:
\describe{
  \item{imagecol}{The x-coordinate of the spot in the image.}
  \item{imagerow}{The y-coordinate of the spot in the image.}
  \item{Spot_ID}{The unique identifier for each spatial spot.}
}
}
\usage{
data(result_local)
}
\description{
A dataset containing spatial coordinates for local analysis, typically used in spatial transcriptomics.
}
\examples{
data(result_local_spatial_coords_df)
head(result_local_spatial_coords_df)
}
\keyword{datasets}
