% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_regional.R
\docType{data}
\name{result_regional}
\alias{result_regional}
\title{result_regional data for SpaCCI}
\format{
A list containing:
\describe{
  \item{pvalue_df}{A data frame of p-values and adjusted p-values for cell-cell interactions.}
}
}
\usage{
data(result_regional)
}
\description{
This example dataset is the result of running the \code{run_SpaCCI} function.
It contains the inferred cell-cell interactions across the regional scale.
}
\details{
These objects can be used for testing and running example analyses with the SpaCCI package.
}
\examples{
  data(result_regional)
  print(result_regional)
}
\keyword{datasets}
