% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test data for SpaCCI}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(test_data)
}
\description{
This dataset includes:
\itemize{
  \item{gene_spt_df}: A data frame of spot-level gene expression.
  \item{cell_prop_df}: A data frame with cell type proportions.
  \item{spatial_coords_df}: A data frame of spatial coordinates.
}
}
\examples{
library(SpaCCI)
data(test_data)
}
\keyword{datasets}
