% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_local.R
\docType{data}
\name{result_local}
\alias{result_local}
\title{result_local data for SpaCCI}
\format{
A list containing:
\describe{
  \item{dataframelist}{A list of data frame of the p-value reults of each spatial neighborhood.}
  \item{RegionIDs_matrix}{A list of matrix contains the spot IDs of each spatial neighborhood.}
}
}
\usage{
data(result_local)
}
\description{
This example dataset is the result of running the \code{run_SpaCCI} function.
It contains the inferred cell-cell interactions across the global scale.
}
\details{
These objects can be used for testing and running example analyses with the SpaCCI package.
}
\examples{
  data(result_local)
  print(result_local)
}
\keyword{datasets}
